remove some tests update some content
