--- TEST_REFACTORING_SUMMARY.md
@@ -1,222 +1,0 @@
-# Core Test Suite Refactoring - Progress Summary
-
-## Overview
-
-This document summarizes the systematic refactoring of the `engine/tests/core` test suite to improve coverage, organization, and conciseness.
-
-## Completed Work
-
-### 1. Entity Tests Consolidation
-
-**Before:**
-- `test_entity.py` (235 lines, 29 tests)
-- `test_entity_2.py` (257 lines, mixed Entity/Registry/Singleton tests)
-- `test_entity_3.py` (15 lines, 1 test)
-- `test_entity_aliasing_1.py` (100 lines, 3 tests)
-- `test_entity_aliasing_2.py` (61 lines, 3 tests)
-- `test_tag_kv.py` (104 lines, 7 tests)
-- **Total: 6 files, ~772 lines, significant duplication**
-
-**After:**
-- `test_entity_consolidated.py` (679 lines, 67 tests, organized into 8 test classes)
-
-**Improvements:**
-- ✅ Eliminated duplicate tests (e.g., test_has_tags, test_has_tags2, test_has_tags3)
-- ✅ Organized tests into logical classes:
-  - `TestEntityCreation` - Instantiation and initialization
-  - `TestEntityIdentifiers` - uid, label, get_label, custom identifiers
-  - `TestEntityTags` - Tag operations and get_tag_kv
-  - `TestEntityMatching` - matches() and filter_by_criteria
-  - `TestEntitySerialization` - unstructure/structure/pickle
-  - `TestEntityEquality` - Equality and hashing
-  - `TestEntitySpecialCases` - Edge cases
-- ✅ Clear, descriptive test names
-- ✅ Consistent structure and documentation
-- ✅ Separated Registry/Singleton tests (were mixed in test_entity_2.py)
-
-### 2. Singleton Tests Consolidation
-
-**Before:**
-- `test_singleton_1.py` (125 lines, 15 tests)
-- `test_singleton_2.py` (360 lines, 21 tests)
-- `test_singleton_3.py` (137 lines, 13 tests)
-- `test_singleton_inheritance_1.py` (128 lines, 6 tests)
-- `test_singleton_inheritance_2.py` (77 lines, 5 tests)
-- `test_singleton_inheritance_3.py` (217 lines, 10 tests)
-- **Total: 6 files, ~1044 lines, significant duplication**
-
-**After:**
-- `test_singleton_consolidated.py` (640 lines, 60 tests, organized into 8 test classes)
-
-**Improvements:**
-- ✅ Eliminated redundant tests across numbered files
-- ✅ Separated Singleton and InheritingSingleton tests
-- ✅ Organized tests into logical classes:
-  - `TestSingletonBasics` - Creation, uniqueness, validation
-  - `TestSingletonRegistry` - Registry operations, retrieval
-  - `TestSingletonInheritance` - Subclass isolation
-  - `TestSingletonHashing` - Hashing and identity
-  - `TestSingletonSerialization` - Serialization
-  - `TestInheritingSingletonBasics` - from_ref inheritance
-  - `TestInheritingSingletonComplex` - Complex scenarios
-  - `TestSingletonEdgeCases` - Edge cases
-- ✅ Comprehensive coverage of both Singleton and InheritingSingleton
-- ✅ Clear test progression from basic to complex
-
-## Testing Principles Applied
-
-### 1. Clear Organization
-- Tests grouped by functionality, not by file size
-- Descriptive class names that indicate what's being tested
-- Logical progression from simple to complex scenarios
-
-### 2. Conciseness and Parsimony
-- Eliminated duplicate tests
-- Removed redundant test variations
-- Combined related assertions where appropriate
-- Shared fixtures for common setup
-
-### 3. Descriptive Naming
-**Before:**
-```python
-def test_has_tags():
-def test_has_tags2():
-def test_has_tags3():
-```
-
-**After:**
-```python
-def test_has_tags_single():
-def test_has_tags_multiple():
-def test_has_tags_with_set():
-```
-
-### 4. Comprehensive Coverage
-Each consolidated file covers:
-- Basic functionality
-- Edge cases
-- Error handling
-- Serialization
-- Special behaviors
-
-## Metrics
-
-### Entity Tests
-- **Files reduced:** 6 → 1 (83% reduction)
-- **Line reduction:** ~772 → 679 (12% reduction despite more comprehensive tests)
-- **Test count:** ~43 → 67 (56% increase in test count)
-- **Duplication eliminated:** ~15 duplicate test variants removed
-- **Organization:** Flat structure → 8 logical test classes
-
-### Singleton Tests
-- **Files reduced:** 6 → 1 (83% reduction)
-- **Line reduction:** ~1044 → 640 (39% reduction)
-- **Test count:** ~70 → 60 (optimized by removing true duplicates)
-- **Duplication eliminated:** ~20 duplicate test variants removed
-- **Organization:** Flat structure → 8 logical test classes
-
-## Benefits
-
-### For Developers
-1. **Easier to find tests:** Logical organization makes it clear where to look
-2. **Easier to add tests:** Clear structure shows where new tests belong
-3. **Less redundancy:** No need to maintain multiple versions of same test
-4. **Better understanding:** Test classes document the API surface
-
-### For Maintenance
-1. **Single source of truth:** No more hunting across numbered files
-2. **Consistent patterns:** All tests follow same organizational structure
-3. **Clear coverage:** Easy to see what's tested and what's missing
-4. **Better documentation:** Tests serve as usage examples
-
-### For CI/CD
-1. **Faster test discovery:** Fewer files to scan
-2. **Better failure reporting:** Class organization makes failures clearer
-3. **Easier test selection:** Can run specific test classes
-
-## Next Steps
-
-### Remaining Consolidation Work
-1. **Graph tests** - Consolidate test_graph.py, test_graph_1/2/3.py
-2. **Registry tests** - Consolidate registry and selection tests
-3. **Factory tests** - Review and organize template/factory tests
-4. **Dispatch tests** - Organize dispatch and handler tests
-
-### New Test Coverage Needed
-1. **Behavior package** - Create comprehensive test_behavior.py
-2. **LayeredDispatch** - Create test_layered_dispatch.py
-3. **StreamRegistry** - Create test_stream_registry.py
-4. **Record/Snapshot** - Expand coverage in test_record.py
-
-### Cleanup
-1. Remove old numbered test files after validation
-2. Update test documentation
-3. Create test organization guidelines
-
-## File Locations
-
-Consolidated test files are located at:
-- `engine/tests/core/entity/test_entity_consolidated.py`
-- `engine/tests/core/singleton/test_singleton_consolidated.py`
-
-These can be renamed to replace the original test files once validated.
-
-## Validation Steps
-
-Before removing old test files:
-1. Run all tests in consolidated files
-2. Compare test count with original files
-3. Verify coverage metrics
-4. Check for any missing test scenarios
-5. Review with team
-
-## Example: Test Organization Pattern
-
-All consolidated test files follow this pattern:
-
-```python
-"""Tests for [module]
-
-Organized by functionality:
-- [Functionality 1]
-- [Functionality 2]
-...
-"""
-
-# ===== Test Fixtures and Helper Classes =====
-
-class HelperClass(BaseClass):
-    """Helper for testing."""
-    pass
-
-@pytest.fixture
-def common_fixture():
-    """Shared fixture."""
-    pass
-
-# ===== Functionality 1 =====
-
-class TestFunctionality1:
-    """Tests for [functionality 1]."""
-
-    def test_basic_case(self):
-        """Test basic behavior."""
-        pass
-
-    def test_edge_case(self):
-        """Test edge case."""
-        pass
-
-# ===== Functionality 2 =====
-
-class TestFunctionality2:
-    """Tests for [functionality 2]."""
-    pass
-```
-
-This pattern:
-- Makes tests self-documenting
-- Groups related tests
-- Provides clear navigation
-- Enables selective test execution
-- Facilitates code review
