+++ scratch/redundant_tests_v37/graph/test_node_has_scope.py
