+++ scratch/redundant_tests_v37/singleton/test_singleton_1.py
