+++ scratch/redundant_tests_v37/singleton/test_singleton_inheritance_1.py
